refactor(tests): organize lexer declaration tests
