+++ crates/rl-tests/tests/lexer/declarations/mod.rs
@@ -1,0 +1,2 @@
+mod constants;
+mod variables;
